Playmind/PlayMindAI/requirements.txt
streamlit==1.33.0
pandas==2.0.3
python-dotenv==1.0.0
fpdf==1.7.2
supabase==2.0.0
openai==1.3.0  # Added as fallback
# IBM Watson - optional, uncomment if you have access
# ibm-watson-machine-learning==1.0.326